pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["LikerK"]
packages = [
  { include = "task_manager" },
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.1"
django-filter = "^21.1"
rollbar = "^0.16.3"
python-dotenv = "^0.20.0"
flake8 = "^4.0.1"

[tool.poetry.dev-dependencies]
gunicorn = "^20.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
